refactor: make interactive-map more componentized
