+++ client/app/features/map/interactive-map/InteractiveMap.tsx
@@ -1,557 +1,28 @@
 import { useState, useRef, useEffect, useCallback } from "react";
 import styles from "./InteractiveMap.module.css";
 import { useNavigationState } from "./useNavigationState";
+import { useTelemetry } from "~/hooks/useTelemetry";
+import type {
+    Point,
+    MapPoint,
+    HazardType,
+    Mode,
+    POIStep,
+    Hazard,
+    DirectionsStep,
+    RoverPosition,
+    ViewBox,
+} from "~/types/map.ts";
 
-// ── Types ──────────────────────────────────────────────
-
-interface Point {
-    x: number;
-    y: number;
-}
-
-export interface MapPoint {
-    id: string;
-    x: number;
-    y: number;
-    label: string;
-    description: string;
-    type: "poi" | "base" | "rover" | "custom";
-}
-
-export type HazardType = "regolith" | "debris" | "crater";
-
-export interface Hazard {
-    id: string;
-    points: Point[];
-    closed: boolean;
-    types: HazardType[];
-}
-
-type Mode = "navigate" | "addPOI" | "plotHazard" | "hazardDetails" | "directions" | "autonomousLTV";
-type POIStep = "placing" | "naming" | "describing";
-type DirectionsStep = "selectDestination" | "review" | "computing" | "active";
-
-export interface RoverPosition {
-    x: number;
-    y: number;
-    heading: number; // degrees, 0 = up/north
-}
-
-interface InteractiveMapProps {
-    // extensibility - add any external props here
-}
-
-interface ViewBox {
-    x: number;
-    y: number;
-    w: number;
-    h: number;
-}
-
-// ── Geometry utilities for pathfinding ─────────────────
-
-function segmentsIntersect(a1: Point, a2: Point, b1: Point, b2: Point): boolean {
-    const d1x = a2.x - a1.x,
-        d1y = a2.y - a1.y;
-    const d2x = b2.x - b1.x,
-        d2y = b2.y - b1.y;
-    const cross = d1x * d2y - d1y * d2x;
-    if (Math.abs(cross) < 1e-10) return false;
-
-    const dx = b1.x - a1.x,
-        dy = b1.y - a1.y;
-    const t = (dx * d2y - dy * d2x) / cross;
-    const u = (dx * d1y - dy * d1x) / cross;
-    return t > 0.001 && t < 0.999 && u > 0.001 && u < 0.999;
-}
-
-function pointInPolygon(p: Point, polygon: Point[]): boolean {
-    let inside = false;
-    for (let i = 0, j = polygon.length - 1; i < polygon.length; j = i++) {
-        const xi = polygon[i].x,
-            yi = polygon[i].y;
-        const xj = polygon[j].x,
-            yj = polygon[j].y;
-        if (yi > p.y !== yj > p.y && p.x < ((xj - xi) * (p.y - yi)) / (yj - yi) + xi) {
-            inside = !inside;
-        }
-    }
-    return inside;
-}
-
-function expandPolygon(polygon: Point[], offset: number): Point[] {
-    // Compute centroid
-    const cx = polygon.reduce((s, p) => s + p.x, 0) / polygon.length;
-    const cy = polygon.reduce((s, p) => s + p.y, 0) / polygon.length;
-    // Push each vertex outward from centroid
-    return polygon.map((p) => {
-        const dx = p.x - cx;
-        const dy = p.y - cy;
-        const dist = Math.sqrt(dx * dx + dy * dy);
-        if (dist === 0) return p;
-        return {
-            x: p.x + (dx / dist) * offset,
-            y: p.y + (dy / dist) * offset,
-        };
-    });
-}
-
-function lineIntersectsPolygon(a: Point, b: Point, polygon: Point[]): boolean {
-    for (let i = 0; i < polygon.length; i++) {
-        const j = (i + 1) % polygon.length;
-        if (segmentsIntersect(a, b, polygon[i], polygon[j])) return true;
-    }
-    return false;
-}
-
-function lineClearsAllHazards(a: Point, b: Point, expandedPolygons: Point[][]): boolean {
-    for (const poly of expandedPolygons) {
-        if (lineIntersectsPolygon(a, b, poly)) return false;
-    }
-    // Also check midpoint isn't inside any polygon
-    const mid = { x: (a.x + b.x) / 2, y: (a.y + b.y) / 2 };
-    for (const poly of expandedPolygons) {
-        if (pointInPolygon(mid, poly)) return false;
-    }
-    return true;
-}
-
-function dist(a: Point, b: Point): number {
-    return Math.sqrt((a.x - b.x) ** 2 + (a.y - b.y) ** 2);
-}
-
-function findPathAroundHazards(start: Point, end: Point, closedHazards: Hazard[]): Point[] {
-    const OFFSET = 15;
-    const expandedPolygons = closedHazards.map((h) => expandPolygon(h.points, OFFSET));
-
-    // If direct line is clear, go straight
-    if (lineClearsAllHazards(start, end, expandedPolygons)) {
-        return [start, end];
-    }
-
-    // Build visibility graph from start, end, and all expanded polygon vertices
-    const nodes: Point[] = [start];
-    const vertexSets: Point[][] = [];
-
-    for (const poly of expandedPolygons) {
-        vertexSets.push(poly);
-        for (const v of poly) {
-            // Only add vertices that aren't inside another polygon
-            let insideAnother = false;
-            for (const otherPoly of expandedPolygons) {
-                if (otherPoly === poly) continue;
-                if (pointInPolygon(v, otherPoly)) {
-                    insideAnother = true;
-                    break;
-                }
-            }
-            if (!insideAnother) nodes.push(v);
-        }
-    }
-    nodes.push(end);
-
-    const n = nodes.length;
-
-    // Build adjacency matrix
-    const adj: number[][] = Array.from({ length: n }, () => Array(n).fill(Infinity));
-    for (let i = 0; i < n; i++) {
-        for (let j = i + 1; j < n; j++) {
-            if (lineClearsAllHazards(nodes[i], nodes[j], expandedPolygons)) {
-                const d = dist(nodes[i], nodes[j]);
-                adj[i][j] = d;
-                adj[j][i] = d;
-            }
-        }
-    }
-
-    // Dijkstra from node 0 (start) to node n-1 (end)
-    const visited = new Array(n).fill(false);
-    const distances = new Array(n).fill(Infinity);
-    const prev = new Array(n).fill(-1);
-    distances[0] = 0;
-
-    for (let iter = 0; iter < n; iter++) {
-        let u = -1;
-        let minDist = Infinity;
-        for (let i = 0; i < n; i++) {
-            if (!visited[i] && distances[i] < minDist) {
-                minDist = distances[i];
-                u = i;
-            }
-        }
-        if (u === -1) break;
-        visited[u] = true;
-
-        for (let v = 0; v < n; v++) {
-            if (!visited[v] && adj[u][v] < Infinity) {
-                const alt = distances[u] + adj[u][v];
-                if (alt < distances[v]) {
-                    distances[v] = alt;
-                    prev[v] = u;
-                }
-            }
-        }
-    }
-
-    // Reconstruct path
-    if (distances[n - 1] === Infinity) {
-        // No path found, fall back to straight line
-        return [start, end];
-    }
-
-    const path: Point[] = [];
-    let cur = n - 1;
-    while (cur !== -1) {
-        path.unshift(nodes[cur]);
-        cur = prev[cur];
-    }
-    return path;
-}
-
-// ── Grid component ─────────────────────────────────────
-
-function Grid({ spacing, viewBox }: { spacing: number; viewBox: ViewBox }) {
-    const lines = [];
-    const startX = Math.floor(viewBox.x / spacing) * spacing;
-    const startY = Math.floor(viewBox.y / spacing) * spacing;
-    const endX = viewBox.x + viewBox.w;
-    const endY = viewBox.y + viewBox.h;
-
-    for (let x = startX; x <= endX; x += spacing) {
-        lines.push(
-            <line
-                key={`v-${x}`}
-                x1={x}
-                y1={viewBox.y}
-                x2={x}
-                y2={viewBox.y + viewBox.h}
-                stroke="#3a3a40"
-                strokeWidth="1"
-            />,
-        );
-    }
-    for (let y = startY; y <= endY; y += spacing) {
-        lines.push(
-            <line
-                key={`h-${y}`}
-                x1={viewBox.x}
-                y1={y}
-                x2={viewBox.x + viewBox.w}
-                y2={y}
-                stroke="#3a3a40"
-                strokeWidth="1"
-            />,
-        );
-    }
-    return <g>{lines}</g>;
-}
-
-// ── Marker component ───────────────────────────────────
-
-function Marker({
-    point,
-    roverPosition,
-    onDelete,
-    isHovered,
-    onHover,
-    onLeave,
-}: {
-    point: MapPoint;
-    roverPosition: RoverPosition;
-    onDelete: (id: string) => void;
-    isHovered: boolean;
-    onHover: () => void;
-    onLeave: () => void;
-}) {
-    const color = "#6ee7b7";
-
-    const dx = point.x - roverPosition.x;
-    const dy = point.y - roverPosition.y;
-    const distance = Math.round(Math.sqrt(dx * dx + dy * dy));
-
-    const descText = point.description?.trim() || "";
-    // Tooltip dimensions — estimate lines based on ~24 chars per line at 11px monospace in 164px width
-    const tooltipW = 180;
-    const charsPerLine = 22;
-    const estimatedLines = Math.ceil(descText.length / charsPerLine);
-    const tooltipH = Math.max(30, estimatedLines * 16 + 14);
-    const tooltipX = point.x - tooltipW / 2;
-    const tooltipY = point.y - 55 - tooltipH;
-
-    return (
-        <g onMouseEnter={onHover} onMouseLeave={onLeave}>
-            {/* Pin shape — tip at (point.x, point.y) */}
-            <path
-                d={`M ${point.x} ${point.y}
-            C ${point.x - 4} ${point.y - 6}, ${point.x - 14} ${point.y - 16}, ${point.x - 14} ${point.y - 26}
-            A 14 14 0 1 1 ${point.x + 14} ${point.y - 26}
-            C ${point.x + 14} ${point.y - 16}, ${point.x + 4} ${point.y - 6}, ${point.x} ${point.y} Z`}
-                fill={color}
-                stroke="#1e1e22"
-                strokeWidth="2"
-            />
-            {/* Inner circle */}
-            <circle cx={point.x} cy={point.y - 26} r="5" fill="#1e1e22" />
-
-            {/* Label */}
-            <text x={point.x} y={point.y + 14} textAnchor="middle" fill="#ccc" fontSize="11">
-                {point.label}: {distance}m
-            </text>
-
-            {/* X delete button — visible on hover */}
-            {isHovered && (
-                <g
-                    cursor="pointer"
-                    onClick={(e) => {
-                        e.stopPropagation();
-                        onDelete(point.id);
-                    }}
-                >
-                    <circle
-                        cx={point.x + 12}
-                        cy={point.y - 38}
-                        r="8"
-                        fill="#1e1e22"
-                        stroke="#888"
-                        strokeWidth="1"
-                    />
-                    <line
-                        x1={point.x + 8}
-                        y1={point.y - 42}
-                        x2={point.x + 16}
-                        y2={point.y - 34}
-                        stroke="#e74c3c"
-                        strokeWidth="2"
-                        strokeLinecap="round"
-                        pointerEvents="none"
-                    />
-                    <line
-                        x1={point.x + 16}
-                        y1={point.y - 42}
-                        x2={point.x + 8}
-                        y2={point.y - 34}
-                        stroke="#e74c3c"
-                        strokeWidth="2"
-                        strokeLinecap="round"
-                        pointerEvents="none"
-                    />
-                </g>
-            )}
-
-            {/* Hover tooltip */}
-            {isHovered && descText && (
-                <g>
-                    <rect
-                        x={tooltipX}
-                        y={tooltipY}
-                        width={tooltipW}
-                        height={tooltipH}
-                        rx="6"
-                        fill="#1e1e22"
-                        stroke="#444"
-                        strokeWidth="1"
-                    />
-                    {/* Arrow */}
-                    <polygon
-                        points={`${point.x - 6},${tooltipY + tooltipH} ${point.x + 6},${tooltipY + tooltipH} ${point.x},${tooltipY + tooltipH + 8}`}
-                        fill="#1e1e22"
-                        stroke="#444"
-                        strokeWidth="1"
-                    />
-                    {/* Cover arrow top border */}
-                    <line
-                        x1={point.x - 6}
-                        y1={tooltipY + tooltipH}
-                        x2={point.x + 6}
-                        y2={tooltipY + tooltipH}
-                        stroke="#1e1e22"
-                        strokeWidth="2"
-                    />
-                    <foreignObject
-                        x={tooltipX + 8}
-                        y={tooltipY + 6}
-                        width={tooltipW - 16}
-                        height={tooltipH - 12}
-                    >
-                        <div
-                            style={{
-                                color: "#ccc",
-                                fontSize: "11px",
-                                lineHeight: "1.4",
-                                overflow: "hidden",
-                                wordWrap: "break-word",
-                                overflowWrap: "break-word",
-                                whiteSpace: "normal",
-                            }}
-                        >
-                            {descText}
-                        </div>
-                    </foreignObject>
-                </g>
-            )}
-        </g>
-    );
-}
-
-// ── Hazard shape component ─────────────────────────────
-
-function HazardShape({ hazard, onDelete }: { hazard: Hazard; onDelete: (id: string) => void }) {
-    const pointsStr = hazard.points.map((p) => `${p.x},${p.y}`).join(" ");
-    const cx = hazard.points.reduce((sum, p) => sum + p.x, 0) / hazard.points.length;
-    const cy = hazard.points.reduce((sum, p) => sum + p.y, 0) / hazard.points.length;
-
-    // Octagon points for warning icon (radius 14, centered at cx,cy)
-    const r = 14;
-    const octagon = Array.from({ length: 8 }, (_, i) => {
-        const angle = (Math.PI * 2 * i) / 8 - Math.PI / 8;
-        return `${cx + r * Math.cos(angle)},${cy + r * Math.sin(angle)}`;
-    }).join(" ");
-
-    return (
-        <g>
-            {hazard.closed ? (
-                <polygon
-                    points={pointsStr}
-                    fill="rgba(255, 138, 117, 0.08)"
-                    stroke="#ff8a75"
-                    strokeWidth="2"
-                />
-            ) : (
-                <polyline points={pointsStr} fill="none" stroke="#ff8a75" strokeWidth="2" />
-            )}
-
-            {hazard.points.map((p, i) => (
-                <circle key={i} cx={p.x} cy={p.y} r="5" fill="#ff8a75" />
-            ))}
-
-            {hazard.closed && (
-                <>
-                    {/* Warning icon at centroid */}
-                    <polygon
-                        points={octagon}
-                        fill="none"
-                        stroke="#ff8a75"
-                        strokeWidth="3"
-                        strokeLinejoin="round"
-                    />
-                    {/* Exclamation mark */}
-                    <line
-                        x1={cx}
-                        y1={cy - 7}
-                        x2={cx}
-                        y2={cy + 2}
-                        stroke="#ff8a75"
-                        strokeWidth="3"
-                        strokeLinecap="round"
-                    />
-                    <circle cx={cx} cy={cy + 7} r="1.5" fill="#ff8a75" />
-
-                    {/* Hazard types label */}
-                    {hazard.types.length > 0 && (
-                        <text x={cx} y={cy + 28} textAnchor="middle" fill="#ff8a75" fontSize="10">
-                            {hazard.types.join(", ")}
-                        </text>
-                    )}
-
-                    {/* X delete button centered below */}
-                    <g
-                        cursor="pointer"
-                        onClick={(e) => {
-                            e.stopPropagation();
-                            onDelete(hazard.id);
-                        }}
-                    >
-                        <circle
-                            cx={cx}
-                            cy={cy + (hazard.types.length > 0 ? 44 : 30)}
-                            r="9"
-                            fill="#1e1e22"
-                            stroke="#888"
-                            strokeWidth="1"
-                        />
-                        <line
-                            x1={cx - 4}
-                            y1={cy + (hazard.types.length > 0 ? 40 : 26)}
-                            x2={cx + 4}
-                            y2={cy + (hazard.types.length > 0 ? 48 : 34)}
-                            stroke="#e74c3c"
-                            strokeWidth="2"
-                            strokeLinecap="round"
-                            pointerEvents="none"
-                        />
-                        <line
-                            x1={cx + 4}
-                            y1={cy + (hazard.types.length > 0 ? 40 : 26)}
-                            x2={cx - 4}
-                            y2={cy + (hazard.types.length > 0 ? 48 : 34)}
-                            stroke="#e74c3c"
-                            strokeWidth="2"
-                            strokeLinecap="round"
-                            pointerEvents="none"
-                        />
-                    </g>
-                </>
-            )}
-        </g>
-    );
-}
-
-// ── Rover icon component ───────────────────────────────
-
-function RoverIcon({ position }: { position: RoverPosition }) {
-    return (
-        <g transform={`translate(${position.x}, ${position.y}) rotate(${position.heading})`}>
-            {/* Shadow */}
-            <ellipse cx="0" cy="2" rx="18" ry="6" fill="rgba(0,0,0,0.3)" />
-
-            {/* Body */}
-            <rect
-                x="-14"
-                y="-12"
-                width="28"
-                height="20"
-                rx="4"
-                fill="#555"
-                stroke="#888"
-                strokeWidth="1.5"
-            />
-
-            {/* Cabin / top */}
-            <rect
-                x="-8"
-                y="-18"
-                width="16"
-                height="10"
-                rx="3"
-                fill="#666"
-                stroke="#888"
-                strokeWidth="1"
-            />
-
-            {/* Wheels */}
-            <circle cx="-16" cy="-8" r="4" fill="#333" stroke="#777" strokeWidth="1" />
-            <circle cx="16" cy="-8" r="4" fill="#333" stroke="#777" strokeWidth="1" />
-            <circle cx="-16" cy="6" r="4" fill="#333" stroke="#777" strokeWidth="1" />
-            <circle cx="16" cy="6" r="4" fill="#333" stroke="#777" strokeWidth="1" />
-
-            {/* Antenna */}
-            <line x1="0" y1="-18" x2="0" y2="-28" stroke="#aaa" strokeWidth="1.5" />
-            <circle cx="0" cy="-29" r="2" fill="#6ee7b7" />
-
-            {/* Direction indicator (front) */}
-            <polygon points="-4,-12 4,-12 0,-16" fill="#6ee7b7" opacity="0.8" />
-
-            {/* Label */}
-            <text x="0" y="22" textAnchor="middle" fill="#6ee7b7" fontSize="10" fontWeight="bold">
-                ROVER
-            </text>
-        </g>
-    );
-}
+import { dist, findPathAroundHazards } from "./Geometry"; // Geometry utilities for pathfinding
+import { Grid } from "./map-components/Grid";
+import { Marker } from "./map-components/Marker"; // Pins
+import { HazardShape } from "./map-components/HazardShape";
+import { RoverIcon } from "./map-components/RoverIcon";
 
 // ── Main map component ─────────────────────────────────
 
-export default function InteractiveMap({}: InteractiveMapProps = {}) {
+export default function InteractiveMap() {
     const svgRef = useRef<SVGSVGElement>(null);
 
     // Only connect to the navigation/stream SSE once the user starts autonomy
@@ -561,14 +32,12 @@ export default function InteractiveMap({}: InteractiveMapProps = {}) {
     const { navState, connected, startAutonomy, stopAutonomy, executePing } =
         useNavigationState(autonomyRequested);
 
-    // Derive rover position from SSE — handle null heading
-    const roverPosition: RoverPosition = navState?.rover_position
-        ? {
-              x: navState.rover_position.x ?? 0,
-              y: navState.rover_position.y ?? 0,
-              heading: navState.rover_position.heading ?? 0,
-          }
-        : { x: 0, y: 0, heading: 0 };
+    const telemetry = useTelemetry();
+    const pos = telemetry.getRoverPosition() ?? { x: 0, y: 0, z: 0 };
+    const roverPosition: RoverPosition = {
+        ...pos,
+        heading: telemetry.getRoverHeading() ?? 0,
+    };
 
     const isAutonomous = navState?.autonomous_driving ?? false;
